refactor: extract shared GPS route types to resolve 4 data clumps
RoutePoint (lat/lng/altitude/speed/timestamp) and the speed stats
trio (max/min/avgSpeedKmh) were duplicated between the rocket-meals
jogging overlay and geonexia's activity storage - reported as 4 data
clumps by data-clumps-doctor.

Both shapes now live in repo-depkit-common (GpsRouteTypes.ts):
- JoggingOverlay's RoutePoint aliases GpsRoutePoint and RunStats
  extends the shared SpeedStats
- ActivityStorage's RoutePoint extends GpsRoutePoint (adding its
  interpolated flag) and SpeedStats re-exports the shared type
