+++ packages/common/index.ts
@@ -24,3 +24,4 @@ export * from './src/AppFeedbackSourceIdentifier';
 export * from './src/CustomerAppStoreIds';
 export * from './src/AppleAppStoreConfig';
 export * from './src/DirectusItemStatus';
+export * from './src/GpsRouteTypes';
